refactor: narrow MCP tool service imports
Import service interfaces in production MCP tool modules from their defining service modules instead of the broad services barrel. This keeps the MCP seam easier to move into a package without changing runtime behavior.
